pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyape"
version = "2.0.0"
description = "EasyApe 🦍 — Text to Stake (Telegram + Discord bot for Bittensor)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "TidalWavesNode"}]
dependencies = [
  "bittensor>=9.5.0",
  "python-telegram-bot>=21.0",
  "discord.py>=2.3.2",
  "pyyaml>=6.0.1",
  "portalocker>=2.8.2",
  "requests>=2.31.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
